Textlog parser integration tests
